Fix: JSON Schema make_nullable modifying values
JSON Schema had a bug in `make_nullable` which was modifying the values passed into it. Since this module aims to be quite functional in nature (due to the static nature of it’s transformations, and the low overhead of operations passed through it) this bug makes building on top of it difficult.