Bump psycopg2 from 2.7.7 -> 2.8.2
Fixes #119

psycopg2 2.8 release introduced mappings between Postgresql errors
and python exceptions.

I assume this came with additional, more granular exceptions that are
now subclassed.

If the original code had been throwing the `IntegrityError` exception, the
code would have still worked, since the `NotNullValidation` is a subclass.
However, since the exception is wrapped in `postgres.PostgresError` and
it's using a matcher, the matching string had to be updated to match the
new exception being thrown.

REFERENCES:
* http://initd.org/psycopg/articles/2019/04/04/psycopg-28-released/
* http://initd.org/psycopg/docs/errors.html