Eagerly flush buffered streams upon receiving a STATE message
In order for orchestrating systems to properly keep track of the STATE of Singer streams, target-postgres needs to emit the STATE messages it receives from the tap to stdout for the orchestrator to persist. This keeps target-postgres Singer-spec compliant and lets users pass that STATE back into the tap to incrementally load from any big sources.

The Singer spec however does not specify what exactly is in STATE messages, so, they can relate to any or all of the active streams. That means that target-postgres doesn't and can't know which records that it may have buffered in memory are "covered" by an incoming STATE record. If target-postgres eagerly emitted the STATE record to stdout, but didn't flush all the records, the system becomes open to inconsistency. If the STATE message were persisted by the outside orchestrator, but then the process crashed later, the buffers in target-postgres process would be lost and data dropped. The implementation prior to this commit suffers from this bug, see https://github.com/datamill-co/target-postgres/pull/120 for more discussion.

This commit prevents this bug by eagerly flushing all the buffered records whenever a STATE message is encountered. When the orchestrator sees a STATE message, it becomes guaranteed safe to persist it as everything seen from the tap has been committed to the database.

The downside of this approach is that it may lead to superfluous flushes of the buffered streams, especially if a tap emits a lot of STATE messages. From my experience with several different taps, they all tend to be written in such a way that they emit STATE messages as a single last step, so I don't think this is too big an issue.

An alternative to this approach would be tracking which records were received before a given STATE message, and then only writing the STATE message out to the orchestrator once all those records had been flushed on their regularly scheduled programming. This means STATE messages are "delayed" until the records they came after have all been flushed on the schedule they would have been flushed on without STATE messages. target-snowflake implements this here: https://gitlab.com/meltano/target-snowflake/blob/master/target_snowflake/target_snowflake.py , but in my mind, this is significant additional complexity for a performance optimization that is in most circumstances unnecessary. It's arguable that writing STATE messages to stdout as soon as they arrive is actually valuable, so that in the presence of a sometimes-failing tap progress can actually be made by getting STATE persisted as soon as possible, and, while the taps are indeed idempotent, rerunning them isn't always free (in terms of API calls and whanot). So, I think this simpler solution would be a great starting point, and after some real world usage we could re-evaluate and move to the more complex approach.