Add the migrations test setup script to the docker entrypoint
The new migrations test requires a couple more virtualenvs to be setup for the test to run correctly. This is handily automated in a setup script, but prior to this change developers had to know the script existed and run it themselves to get their development environment to run the tests successfully.

This changes the development environment setup script to run the migrations test setup script as well so the tests should run cleanly from the very beginning of A Developer's Journey.