Emit STATE messages if only one of many registered streams is adding records
This fixes a small but annoying bug in the STATE message support for taps that output records to many streams at once, and especially taps that emit a whole whack of schemas at the start of the run, and then emit all the records for each stream one by one.

Before this change, the safe-to-emit-STATE-message watermark was the global minimum of each stream's most recently flushed watermark. This works well when you have many streams emitting records at once due to the whole "don't know which streams each STATE message covers so we have to be sure all records from before the STATE message was emitted have been flushed before passing along the STATE message" thing. The add watermark rises higher than the flushed watermark as records are buffered, and then the flush watermark catches up when a flush happens, moving the global minimum up and permitting new STATE messages to be emitted.

Before this change, upon receiving a SCHEMA message for a new stream, the stream is registered, and a most recently added and most recently flushed watermark of 0 is set for that stream. That makes sense because no records for that stream have been added yet, so no records have been flushed yet, so the flush watermark is set to the same thing that the add watermark is. All the SCHEMA message sets up is an expectation for what records might (or might not ever) come. For the purposes of emitting STATE messages, the SCHEMA message doesn't really hold any value, as STATE is really only relative to records.

So, the issue is that the previous version of this code initialized each stream's most recently flushed watermark to 0 upon receiving the SCHEMA message, which locked the global minimum flush watermark at 0 until any records for that stream arrived. If the tap works in such a way that it moves through each stream making API calls for that stream, this means that until the very last stream is processed, that stream's flush watermark at 0, as no records have been flushed yet, but that also means the global flush watermark is at 0 too. This leads to no state messages being output until the end. D'oh.

This changes things around in the tracker class to not include streams who haven't yet seen records to not be included in the global threshold calculation. This has the effect of allowing straightforward initialization on SCHEMA messages while allowing streams to be processed incrementally and STATE messages to be emitted without including not-yet-in-use streams. The code to do this is pretty verbose because the issues are subtle, so I've tried to comment a bit more than usual and be explicit.

I'm a little concerned about this whole thing as I think this is beginning to violate the KISS rule. I try to abide by the whole "write code only as half as smart as you are because you need to be twice as smart as the code to debug it", and, based on the amount of issues coming out of this, it would appear I am at the limits of my smartness. If anyone has suggestions for a simpler approach or a better testing strategy I am all ears!