Document the config params that govern row batching (#148)
`max_batch_rows`, `max_batch_size`, and `batch_detection_threshold` were all config properties that existed prior to this change and are threaded through into the BufferedSingerStream to configure it's batching behaviour. This adds docs for those params to the README.

In addition, this makes the default for the `batch_detection_threshold` configuration value dynamic so that it adjusts to be proportionally the same relative to the `max_batch_rows` if `max_batch_rows` is adjusted and `batch_detection_threshold` is unset. I think it would be surpising for a user if they adjusted `max_batch_rows` to something low like `1000` only to find that the detection check was made by default every `5000` rows, giving an effective `max_batch_rows` of `5000`. The value is still configurable as it has always been, but after this change, if `max_batch_rows` is set, `target_postgres` infers a reasonable value for the batch detection threshold by taking the max of (1/40 * `max_batch_rows`) and `50`.