Make batch writing the StreamTracker's responsibility for clarity and simplicity
Before this change, it was a little confusing exactly who and when the write_batch method should be called, but since all batch flushing is now done through the StreamTracker, we can make batch writing exclusively its responsibility. It was only done in one other place (ACTIVATE_VERSION messages), but now that it's all in this place, the flush_stream and flush_streams code can get a bit simpler. This also renames _flush_stream to something a bit more descriptive and removes one condition that wasn't actually necessary.