Add an initial_sql config option to execute arbitrary SQL upon boot (#136)
* Add an initial_sql config option to execute arbitrary SQL upon boot

This lets users execute some SQL before target-postgres spins up and does it's thing. This is useful for SQL like `SET ROLE` or setting up encoding business, and I think will end up being generally useful for other stuff I haven't (and don't need to) predict.

For me, this was prompted by needing to `SET ROLE` on Google Cloud SQL. On some postgres deployments, target-postgres may be running as a user with limited permissions, or as a user different than the downstream consumer of the data. In these instances it's important that the table owner of the target-postgres created tables is controllable, especially in deployments where the superuser isn't in the developers control, like Google Cloud SQL. `SET ROLE` is the only way to really control the table owner of the created tables so that other users may be able to grant permissions on those tables outside of the normal `target-postgres` lifecycle. See https://stackoverflow.com/questions/2276644/list-all-tables-in-postgresql-information-schema for a more thourough description of the pieces at play here.

* Rename initial_sql to before_run_sql and add after_run_sql SQL hooks